more COLRv1 bug fixes

+++ Harfbuzz/Structs/Font.cs
@@ -38,9 +38,9 @@ namespace TextMeshDOTS.HarfBuzz
         {
             HB.hb_font_draw_glyph(ptr, glyphID, drawFunctions, ref drawData);
         }
-        public void PaintGlyph(uint glyphID, ref PaintData paintData, PaintDelegates paintFunctions, uint palette,ColorARGB foreground)
+        public void PaintGlyph(uint glyphID, ref PaintData paintData, PaintDelegates paintFunctions, uint palette, ColorARGB foreground)
         {
-            HB.hb_font_paint_glyph(ptr, glyphID, paintFunctions, ref paintData, palette, foreground);
+            HB.hb_font_paint_glyph(ptr, glyphID, paintFunctions, ref paintData, palette, (uint)foreground);
         }
 
         public void GetSyntheticBold(out float x_embolden, out float y_embolden, out bool in_place)
